Move state to separate module
